requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
hdbscan==0.8.40
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.16
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.161
ultralytics-thop==2.0.14
umap-learn==0.5.8
urllib3==2.5.0
yt-dlp==2025.6.30
